LPS-65012 Document library move constant class back and export the package
