+++ modules/apps/collaboration/document-library/document-library-web/src/main/resources/com/liferay/document/library/web/constants/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
